Renamed Examples folders to CodeConcepts. Added DLL example, more questions, UML PacketManager.
Renamed Example CodeConcepts so it was a bit more clear.  I will be
working on an Examples folder later which will use the CommProtocol
library rather than just general C++ programming.  I added PacketManager
and all of the classes it owns to the UML as well as outlining the
relationships of the other objects in the library.  Also, there are a
lot more questions!